[14.0][FIX] base_tier_validation, change difinition should not affect reviews
